Load favicons from popup after form submission has completed
This means we will no longer attempt to associate an entry with the favicon
that was available at the time the form was submitted. In most cases this
makes no difference but the more accurate previous approach would fail
intermittently for reasons that are unavoidable since WebExtensions and
e10s.