Re-organise page.ts to tolerate insertion into non-pages
Browsers sometimes inject content scripts into SVG images or other
document types that are not able to contain forms.

Since that behaviour is undocumented, I'm not sure if it is a browser
bug or just something extension developers are supposed to not
worry about. In Kee's case, this causes unnecessary work and can
even crash the browser, although the mechanism for this is not
entirely clear.