Remove old workaround for unreliable port connections
It is possible that some slow pages or computers could do too much
work or fail to start Kee on the affected web page.

I hypothesise that the underlying web browser bugs which
necessitated this workaround have since been fixed but they were
undocumented so we must test in the real world to be sure that this
doesn't re-introduce an old set of problems.